backend/requirements/base.txt
# ========================
# Web Framework
# ========================
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0

# ========================
# Database
# ========================
psycopg2-binary==2.9.9

# ========================
# Task Queue & Caching
# ========================
celery==5.4.0
redis==5.0.7
django-redis==5.4.0
django-celery-beat==2.6.0
django-celery-results==2.5.1

# ========================
# Scraping
# ========================
beautifulsoup4==4.12.3
playwright==1.45.0
lxml==5.2.2
httpx==0.27.0

# ========================
# Utilitas
# ========================
python-decouple==3.8
python-dateutil==2.9.0

# ========================
# Monitoring
# ========================
sentry-sdk[django]==2.10.0

# ========================
# Validasi & Filter
# ========================
django-filter==24.2
playwright-stealth==1.0.6
setuptools>=70.0.0